+++ components/ListPostsEdit/index.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
-import { usid } from 'utils';
-import styles from './ListPostsEdit.module.css';
+import { uid } from 'uid/secure';
+import styles from './ListPostsEdit.module.scss';
 import querys from "utils/querys";
 import { postMockup } from 'utils/mockups';
 import Link from 'next/link';
@@ -22,7 +22,6 @@ const ListPostsEdit = () => {
           ...data[key]
         })
       }
-      console.log('init set',list)
       setPosts(list)
     }
 
@@ -39,14 +38,17 @@ const ListPostsEdit = () => {
       for(let key in posts){
         list.push(posts[key])
       }
-      console.log('update set',posts)
       setPosts(list)
     }
   }
     
   const onDeletePostBtn = async (id) => {
-    await querys.deletePost(id)
-    updatePosts()
+    const needDeleting = confirm('Точно удалить пост?')
+
+    if(needDeleting) {
+      await querys.deletePost(id)
+      updatePosts()
+    }
   }
 
   const onUpdatePostBtn = async (id, title) => {
@@ -55,7 +57,7 @@ const ListPostsEdit = () => {
   }
 
   const onAddPostBtn = async () => {
-    const newId = usid.uuid()
+    const newId = uid(14)
 
     const post = {
       ...postMockup,
@@ -74,17 +76,17 @@ const ListPostsEdit = () => {
       
     return (
       <div className={styles.onePost}>
+        <div className={styles.editBtn}>
+          <Link href={`/admin/post/${id}`} >
+            edit
+          </Link>
+        </div>
         <div className={styles.id}>{id}</div>
         <div className={styles.title}>
           <input type="text" value={inputVal} onChange={(e) => {setInputVal(e.target.value)}} />
         </div>
         <div className={styles.deleteBtn} onClick={() => onDeletePostBtn(id)}>delete</div>
         <div className={styles.updateBtn} onClick={() => onUpdatePostBtn(id, inputVal)}>update</div>
-        <div className={styles.editBtn}>
-          <Link href={`/admin/post/${id}`} >
-            edit
-          </Link>
-        </div>
       </div>
     )
   }
